feat(STONEINTG-399): move cache functions to separate package
* Update controllers to use cache package
* Update unit tests to use cache package
* Rename binding index functions to reflect usage
