+++ helpers/helpers_suite_test.go
@@ -15,12 +15,11 @@ package helpers_test
 
 import (
 	"context"
+	"github.com/redhat-appstudio/integration-service/cache"
 	goodies "github.com/redhat-appstudio/operator-goodies/test"
 	"go/build"
 	"path/filepath"
 	"testing"
-
-	"github.com/redhat-appstudio/integration-service/controllers/pipeline"
 
 	"k8s.io/client-go/rest"
 
@@ -92,7 +91,7 @@ var _ = BeforeSuite(func() {
 	k8sClient = k8sManager.GetClient()
 	go func() {
 		defer GinkgoRecover()
-		Expect(pipeline.SetupIntegrationTestScenarioCache(k8sManager)).To(Succeed())
+		Expect(cache.SetupIntegrationTestScenarioCache(k8sManager)).To(Succeed())
 		Expect(k8sManager.Start(ctx)).To(Succeed())
 	}()
 })
