+++ release/release_suite_test.go
@@ -15,6 +15,7 @@ limitations under the License.
 
 import (
 	"context"
+	"github.com/redhat-appstudio/integration-service/cache"
 	goodies "github.com/redhat-appstudio/operator-goodies/test"
 	"go/build"
 	"path/filepath"
@@ -31,7 +32,6 @@ import (
 
 	applicationapiv1alpha1 "github.com/redhat-appstudio/application-api/api/v1alpha1"
 	integrationalpha1 "github.com/redhat-appstudio/integration-service/api/v1alpha1"
-	"github.com/redhat-appstudio/integration-service/controllers/snapshot"
 	releasev1alpha1 "github.com/redhat-appstudio/release-service/api/v1alpha1"
 	clientsetscheme "k8s.io/client-go/kubernetes/scheme"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -93,7 +93,7 @@ var _ = BeforeSuite(func() {
 	k8sClient = k8sManager.GetClient()
 	go func() {
 		defer GinkgoRecover()
-		Expect(snapshot.SetupReleasePlanCache(k8sManager)).To(Succeed())
+		Expect(cache.SetupReleasePlanCache(k8sManager)).To(Succeed())
 		Expect(k8sManager.Start(ctx)).To(Succeed())
 	}()
 })
